moved AddFlight validators to separate folder
