+++ FlightPlanner.Services/Validators/AddFlight/AddFlightRequestValidator.cs
@@ -1,7 +1,7 @@
 ﻿using FlightPlanner.Core.DTO;
 using FlightPlanner.Core.Services;
 
-namespace FlightPlanner.Services.Validators
+namespace FlightPlanner.Services.Validators.AddFlight
 {
     public class AddFlightRequestValidator : IValidator
     {
