+++ FlightPlanner.Services/Validators/AddFlight/CarrierValidator.cs
@@ -1,0 +1,13 @@
+﻿using FlightPlanner.Core.DTO;
+using FlightPlanner.Core.Services;
+
+namespace FlightPlanner.Services.Validators.AddFlight
+{
+    public class CarrierValidator : IValidator
+    {
+        public bool IsValid(AddFlightDto dto)
+        {
+            return !string.IsNullOrEmpty(dto?.Carrier);
+        }
+    }
+}
