+++ FlightPlanner.Services/Validators/AddFlight/TimeFrameValidator.cs
@@ -1,8 +1,8 @@
-﻿using FlightPlanner.Core.DTO;
+﻿using System;
+using FlightPlanner.Core.DTO;
 using FlightPlanner.Core.Services;
-using System;
 
-namespace FlightPlanner.Services.Validators
+namespace FlightPlanner.Services.Validators.AddFlight
 {
     public class TimeFrameValidator : IValidator
     {
