+++ FlightPlanner.Services/Validators/AddFlight/ToAirportValidator.cs
@@ -1,13 +1,13 @@
 ﻿using FlightPlanner.Core.DTO;
 using FlightPlanner.Core.Services;
 
-namespace FlightPlanner.Services.Validators
+namespace FlightPlanner.Services.Validators.AddFlight
 {
-    public class CarrierValidator : IValidator
+    public class ToAirportValidator : IValidator
     {
         public bool IsValid(AddFlightDto dto)
         {
-            return !string.IsNullOrEmpty(dto?.Carrier);
+            return dto?.To != null;
         }
     }
 }
